moved db components
