+++ public-ui/src/app/data-browser/data-browser.module.ts
@@ -7,13 +7,13 @@ import { HighchartsStatic } from 'angular2-highcharts/dist/HighchartsService';
 import * as highcharts from 'highcharts';
 import 'highcharts/highcharts-more';
 import { DbNoResultsComponent } from '../components/db-no-results/db-no-results.component';
-import { RecursiveTreeComponent } from '../components/recursive-tree/recursive-tree.component';
-import { SourceTreeComponent } from '../components/source-tree/source-tree.component';
 import { SharedModule } from '../shared/shared.module';
 import { ChartComponent } from './chart/chart.component';
 import { ChartBaseComponent } from './charts/chart-base/chart-base.component';
 import { ChartTopResultsComponent } from './charts/chart-top-results/chart-top-results.component';
 import { ChartService } from './charts/chart.service';
+import { RecursiveTreeComponent } from './components/recursive-tree/recursive-tree.component';
+import { SourceTreeComponent } from './components/source-tree/source-tree.component';
 import { ConceptChartsComponent } from './concept-charts/concept-charts.component';
 import { DataBrowserRoutingModule } from './databrowser-routing.module';
 import { EhrViewComponent } from './views/ehr-view/ehr-view.component';
